+++ packages/persistence-adapters/indexeddb/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* Added support for custom database prefix option (thanks to @pierissimo!)
+
 ## [1.0.0] - 2025-01-06
 
 * Initial release